Court Marriage Summary
This project analyzes court marriages from September of 2018 to May of 2022 in Nigeria. It highlights the different services provided by all the approved marriage registries, the service trends, funds generated from each service and the status. I created 2 similar visuals displaying the same information. A video analysis of this project is on my linkedin page, feel free to check it out